CU-869ahw0mw: Add argument to control data flow when saving results. (#144)
* CU-869ahw0mw: Add argument to control data flow when saving results.

When  is provided, the user (probably) expects the data to be saved on disk upon method call.
But the current implementation forced the user to iterate over the results to force the annotation to actually happen.
So this change allows the method to materialise the list internally to force the annotation to happen and results to be saved on disk.
Additionally, it adds 2 other options:
1. The lazy iteration (what happens when no  is provided) where the iteration of data is left to the user
2. The combined / saved and return option where the results are materialised, but also yielded. Notably, this will take up a lot of memory if/when used with large data sets

* CU-869ahw0mw: Make tests run without materialising the output for multiprocessing

* CU-869ahw0mw: Move DeID tests to non-deprecated method

* CU-869ahw0mw: Some whitespace fixes

* CU-869ahw0mw: Fix issue withe multiprocessing.

The previous implementation would always consider the method a generator. And as such, the work would never be done
at call time, regardless of whether or not the  was provided. This commit fixes that by making the
wrapper method a regular method that (sometimes) returns the iterator and other times just a (potentially empty) list.

* CU-869ahw0mw: Add further tests to new functionality

* CU-869ahw0mw: Fix behaviour (so it remains the same) in old test

* CU-869ahw0mw: Fix test regarding generator issue

* CU-869ahw0mw: Move saving (and not returning data) to a separate method

* CU-869ahw0mw: Update tests accordingly as per last change